docs(adl): Use Hexagonal (="Clean") Architecture
I use the term "hexagonal architecture" over "clean architecture" in
order to avoid the ambuiguity tied to the existence of the book "Clean
Architecture" from Uncle Bob that is about Architecture with a big A,
i.e. tradeoffs, good practices and smells. The "hexagonal architecture",
a synonym (in my current understanding) of "clean architecture", is on
the other hand already a position taken on some of these tradeoffs